+++ app/[locale]/projects/[projectId]/dmp/create/page.tsx
@@ -50,6 +50,7 @@ interface PublicTemplatesInterface {
   name?: string | null;
   description?: string | null;
   modified?: string | null;
+  modifiedByName?: string | null;
   modifiedById?: number | null;
   visibility?: string | null;
   ownerDisplayName?: string | null;
@@ -183,6 +184,7 @@ const PlanCreate: React.FC = () => {
 
   // Transform the templates data into more useable format
   const transformTemplates = async (templates: (PublicTemplatesInterface | null)[]) => {
+    console.log('Transforming templates:', templates);
     const transformedTemplates = await Promise.all(
       templates.map(async (template: PublicTemplatesInterface | null) => ({
         id: template?.id,
@@ -200,7 +202,7 @@ const PlanCreate: React.FC = () => {
         funder: template?.ownerDisplayName || "",
         funderUri: template?.ownerURI || "",
         lastUpdated: template?.modified ? formatDate(template?.modified) : null,
-        lastRevisedBy: template?.modifiedById || null,
+        lastRevisedBy: template?.modifiedByName || null,
         hasAdditionalGuidance: false,
         defaultExpanded: false,
         visibility: template?.visibility,